bta: le_audio: Add testAutonomousReleaseFromEnablingState test case
This adds testAutonomousReleaseFromEnablingState test case.
The scenario checks if the stream configuration is not aborted
when one of the devices being in Enabling state is autonomously
released and the PACS Available Contexts are cleared.
The test checks if the device can be configured later on
once the available contexts are back.

Bug: 357472821
Flag: Exempt; TEST_ONLY
Test: atest bluetooth_le_audio_test
Change-Id: Ie32f0678c2d732745a6acde381c9f2c0718c8c86