Bass: Remove unused BASE consolidation
The consolidated BASE content is never used, and additionally it is
executed without checking if the codec configuration uses LTV format.
If the codec is using a custom configuration format, the consolidation
code misinterpret it as LTVs and tries to consolidate as LTVs.
The consolidated configs and metadata is never used and can be safely
removed.

Bug: 328181940
Test: atest BluetoothInstrumentationTests
Flag: EXEMPT; Removes dead code
Change-Id: I40034f95a10924c770b84d7f0df985abbf20993e