floss: check if BCC is allowed to open SCO
In HFPv1.9, 4.11.2, Audio Connection Setup by HF:

> When the HF triggers the establishment of the Codec Connection
it shall send the AT command AT+BCC to the AG. The AG shall respond
with OK if it will start the Codec Connection procedure, and with ERROR
if it cannot start the Codec Connection procedure.

In CrOS, SCO initiated from the peer is not a supported use case. In
this CL we introduce a check to reject the BCC open request.

Bug: 353234804
Flag: EXEMPT floss-only
Test: m Bluetooth
Change-Id: I339c68c537d842fcc6b71ce17e7a35389abbc043