Queue DIS requests
Instead of rejecting a DIS request when there's already another
ongoing DIS request, just queue it and carry it out as soon as the
first request is done.

Additionally, don't proceed to connect GATT if we're not currently
connected, since we have no handler for both cases where it is failed
and successful. Since we already marked the request as ongoing, any
attempt to connect GATT is guaranteed to block all future DIS requests,
regardless of whether we queue the DIS requests or not.

Bug: 357011645
Bug: 360289803
Test: m -j
Test: Purposely modify the code to invoke tons of DIS request. Verify
      the requests are all carried out.
Flag: com.android.bluetooth.flags.queue_dis_requests

Change-Id: I5e679f99296625a770c45268914262d3d9f48664