leaudio: Fix invalid state machine timeout
This patch fixes following scenario
1. State machine is ordered to start the stream and Codec Config CMD
is send out to the remote, but there is not yet answer

2. Group is set to inactive while waiting for CODEC CONFIGURED state.

In this particular case, internal state machine watchdog was not cleared
which lead later to disconnect devices.

THis is fixed here.

Bug: 361009375
Test: atest bluetooth_le_audio_test bluetooth_le_audio_client_test
Flag: Exempt, regression tested with unit tests, tests checking new
behaviour has been added

Change-Id: I5414534c56f170f836b6a6b7dd894d9252bdd8ee