Update test node constructor to use random UUID
BrowseNode has a device based constructor, which is used by tests to
create nodes. Its not current used anywhere for real during AVRCP
browsing with real devices (the state machine doesn't call the
onConnected() function that calls it in BrowseTree). This constructor
should still use the UUID based method for creating a media id.

In the future, we can use this constructor directly in code as well.

Bug: 332367017
Bug: 363009246
Test: atest BrowseNodeTest.java
Change-Id: I78818607bb3c32dd684cd4a4d3ea95e67fc53fe6