Errorprone: JavaUtilDate
Date has a bad API that leads to bugs;
prefer java.time.Instant or LocalDate.

Since it is sometimes complicated to change this, this CL aim to enforce
the Errorprone for futur usage, but let the current one and only
suppress the error for existing issues.
Usage of these method should not be extended

Test: m Bluetooth
Bug: 365629730 # to track removal of exception
Bug: 344658662
Flag: Exempt, enforce error in the futur
Change-Id: I150d1a5ac63dfb756fa5711afc8dd9311ecfb279