Use SDP size in next byte for small Report descriptors
When the length of the Report descriptor for HIDD is below 256 bytes but the length in its SDP attribute is specified with SIZE_IN_NEXT_WORD, MS Windows no longer correctly accepts the HIDD connection.

To retain compatibility with existing apps and Windows releases, we can use SIZE_IN_NEXT_BYTE if possible based on the descriptor size, which works around the issue.

Bug: 362791956
Test: manually
Change-Id: I5f13aa571cd373876d1cf94c3ac4f7d5facd9a53