Mock TelephonyManager in ProfileService tests
When starting PBAP Service, it eventually tries to use TelephonyManager.
This current tries to use the _real_ implementation of the
function, since our AdapterService object uses the device under test's
context and doesn't mock everything on its own. This implementation can
throw exceptions based on the underlying configuration of the device.

We want to mock TelephonyManager so we can ensure calls to it in tests
always work and don't break our tests.

Bug: 366622623
Flag: EXEMPT, test only change
Test: atest com.android.bluetooth.btservice.ProfileServiceTest
Change-Id: If80ac4a5df7074bcec9c7fee8c149219f09536af