Ignore duplicate remove device request
A "pending remove" entry is created when the removeBond() is requested
but the device is not disconnected. If another removeBond() is requested
for the same device before first one is concluded, another entry is
created. As all the transports are disconnected, the first entry is
serviced. But the second entry may remain their preventing further
connections to the same device.
This change ignores the duplicate remove device request.

Test: mmm packages/modules/Bluetooth
Bug: 339784340
Flag: EXEMPT bugfix
Change-Id: I73747684f8eab2cc4a016d8cd458c00b9c672f2c