Floss: Increase the cleanup timeout on ChromeOS
When the kernel executes close() on a HCI socket, it calls a
vendor-specific shutdown routine. Some vendors implement this routine by
first performing an HCI reset and waiting up to 10 seconds. This delay
means that stopping the HciHalHost module requires a timeout of at least
10 seconds to prevent premature cleanup due to the timeout, which could
lead to an unexpected sequence of operations and ultimately a crash.

This patch increases the cleanup timeout to 11 seconds on ChromeOS and
enabling other platforms to set their own value.

Bug: 365936141
Tag: #floss
Test: mmm packages/modules/Bluetooth
Test: modify the kernel to create a delay in close socket and observe no
crash on disabling the adapter on ChromeOS.
Flag: EXEMPT, Floss-only changes

Change-Id: I81bfa791e594f1b1c747bb9bcfa32dd2e281f822