floss: don't report properties_changed on device_found
In the UI layer, we treat the UUID updates from properties_changed
event as completion of service discovery, then we forward it to ARC
as well. While this is not ideal, there is no easy way even in rust
layer to know when exactly the service discovery is done.

Starting with aosp/3189612, we also call properties_changed in
device_found. Although it's logical to report the new properties of
a newly found device, this contradicts the prior assumption of
service discovery completion, and messes up with ARC interaction.

This reverts a part of aosp/3189612 for a simple workaround. Later we
need to solve the root cause of the issue, which is the lack of
knowledge when the service discovery is actually done.

Bug: 367255907
Test: m -j
Test: CTSV LE Secure Connection
Flag: EXEMPT, floss-only changes
Change-Id: Ic7d4028fc42cb16b5ec8ed0b28335fd361253a31